Fixed viewer and code cleanup
