+++ src/lib/modular_slam/include/modular_slam/backend_interface.hpp
@@ -1,30 +1,24 @@
 #ifndef MSLAM_BACKEND_INTERFACE_H_
 #define MSLAM_BACKEND_INTERFACE_H_
 
-#include "modular_slam/constraints_interface.hpp"
-#include "modular_slam/frontend_interface.hpp"
+#include "modular_slam/frontend_output.hpp"
 #include "modular_slam/keyframe.hpp"
 #include "modular_slam/landmark.hpp"
 #include "modular_slam/slam3d_types.hpp"
 #include "modular_slam/slam_component.hpp"
+#include <unordered_set>
 
 namespace mslam
 {
-
-template <typename LandmarkStateType>
-struct MergedLandmark
-{
-};
 
 template <typename SensorStateType, typename LandmarkStateType, typename ObservationType>
 struct BackendOutput
 {
     using LandmarkType = Landmark<LandmarkStateType>;
     using KeyframeType = Keyframe<SensorStateType>;
-    using MergedLandmarkType = MergedLandmark<LandmarkStateType>;
 
-    std::vector<std::shared_ptr<LandmarkType>> updatedLandmarks;
-    std::vector<std::shared_ptr<KeyframeType>> updatedKeyframes;
+    std::unordered_set<std::shared_ptr<LandmarkType>> updatedLandmarks;
+    std::unordered_set<std::shared_ptr<KeyframeType>> updatedKeyframes;
     std::vector<LandmarkKeyframeObservation<SensorStateType, LandmarkStateType, ObservationType>> outlierObservations;
 };
