+++ src/lib/modular_slam/include/modular_slam/frontend_interface.hpp
@@ -1,7 +1,7 @@
 #ifndef MSLAM_FRONTEND_INTERFACE_HPP_
 #define MSLAM_FRONTEND_INTERFACE_HPP_
 
-#include "modular_slam/constraints_interface.hpp"
+#include "modular_slam/frontend_output.hpp"
 #include "modular_slam/landmark.hpp"
 #include "modular_slam/slam_component.hpp"
 #include <any>
@@ -9,27 +9,6 @@
 
 namespace mslam
 {
-
-template <typename SensorDataType>
-struct LoopDetection
-{
-    std::shared_ptr<Keyframe<SensorDataType>> firstKeyframe;
-    std::shared_ptr<Keyframe<SensorDataType>> secondKeyframe;
-};
-
-template <typename SensorStateType, typename LandmarkStateType, typename ObservationType>
-struct FrontendOutput
-{
-    std::optional<LoopDetection<SensorStateType>> loopDetected;
-    std::vector<LandmarkObservation<LandmarkStateType, ObservationType>> landmarkObservations;
-    std::shared_ptr<Keyframe<SensorStateType>> newKeyframe;
-    std::vector<std::shared_ptr<Landmark<LandmarkStateType>>> newLandmarks;
-
-    // TODO: update with traits!
-    // TODO 2: it would be more efficient to split keypoint / descriptor data
-    SensorStateType sensorState;
-    // SensorStateType slamState;
-};
 
 template <typename SensorDataType, typename SensorStateType, typename LandmarkStateType, typename ObservationType>
 class FrontendInterface : public SlamComponent
